# api/main.py
from fastapi import FastAPI, Depends, HTTPException
from fastapi.security import OAuth2PasswordBearer
from pydantic import BaseModel
from typing import List, Optional

app = FastAPI()

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

# Models
class User(BaseModel):
    username: str
    password: str

class Metric(BaseModel):
    cpu_usage: int
    memory_usage: int
    instance_id: str

class Prediction(BaseModel):
    predicted_issue: str
    confidence: float
    predicted_actions: List[str]

class ScalingLog(BaseModel):
    action_taken: str
    reason: str
    instance_ids: List[str]

# Authentication Endpoints
@app.post("/auth/register", tags=["Authentication"], status_code=201)
async def register(user: User):
    # Implement user registration logic
    return {"message": "User registered"}

@app.post("/auth/login", tags=["Authentication"], status_code=200)
async def login(user: User):
    # Implement user login logic
    return {"message": "User logged in", "token": "JWT token"}

@app.post("/auth/logout", tags=["Authentication"], status_code=204)
async def logout(token: str = Depends(oauth2_scheme)):
    # Implement user logout logic
    return {"message": "User logged out"}

# Metrics Endpoints
@app.get("/api/metrics", tags=["Metrics"], status_code=200)
async def get_metrics(token: str = Depends(oauth2_scheme)):
    # Implement logic to get metrics
    return [{"timestamp": "2023-10-01T00:00:00Z", "cpu_usage": 50, "memory_usage": 1024, "instance_id": "i-123456"}]

@app.post("/api/metrics", tags=["Metrics"], status_code=201)
async def add_metric(metric: Metric, token: str = Depends(oauth2_scheme)):
    # Implement logic to add a new metric
    return {"message": "Metric added"}

@app.get("/api/metrics/{id}", tags=["Metrics"], status_code=200)
async def get_metric(id: str, token: str = Depends(oauth2_scheme)):
    # Implement logic to get a specific metric
    return {"timestamp": "2023-10-01T00:00:00Z", "cpu_usage": 50, "memory_usage": 1024, "instance_id": "i-123456"}

@app.put("/api/metrics/{id}", tags=["Metrics"], status_code=200)
async def update_metric(id: str, metric: Metric, token: str = Depends(oauth2_scheme)):
    # Implement logic to update a specific metric
    return {"message": "Metric updated"}

@app.delete("/api/metrics/{id}", tags=["Metrics"], status_code=204)
async def delete_metric(id: str, token: str = Depends(oauth2_scheme)):
    # Implement logic to delete a specific metric
    return {"message": "Metric deleted"}

# AI Predictions Endpoints
@app.get("/api/predictions", tags=["AI Predictions"], status_code=200)
async def get_predictions(token: str = Depends(oauth2_scheme)):
    # Implement logic to get AI predictions
    return [{"predicted_issue": "Issue A", "confidence": 0.95, "predicted_actions": ["Action 1", "Action 2"]}]

@app.post("/api/predictions", tags=["AI Predictions"], status_code=201)
async def add_prediction(prediction: Prediction, token: str = Depends(oauth2_scheme)):
    # Implement logic to add a new AI prediction
    return {"message": "AI Prediction added"}

# Scaling Logs Endpoints
@app.get("/api/scaling_logs", tags=["Scaling Logs"], status_code=200)
async def get_scaling_logs(token: str = Depends(oauth2_scheme)):
    # Implement logic to get scaling logs
    return [{"action_taken": "Scale up", "reason": "High CPU usage", "instance_ids": ["i-123456", "i-789012"]}]

@app.post("/api/scaling_logs", tags=["Scaling Logs"], status_code=201)
async def add_scaling_log(log: ScalingLog, token: str = Depends(oauth2_scheme)):
    # Implement logic to add a new scaling log
    return {"message": "Scaling log added"}

# Frontend Endpoints
@app.get("/api/dashboard/summary", tags=["Frontend"], status_code=200)
async def get_dashboard_summary(token: str = Depends(oauth2_scheme)):
    # Implement logic to get dashboard summary
    return {"summary": "Dashboard summary"}

@app.get("/api/dashboard/metrics/trend", tags=["Frontend"], status_code=200)
async def get_metrics_trend(token: str = Depends(oauth2_scheme)):
    # Implement logic to get metrics trend
    return {"trend": "Metrics trend"}

@app.get("/api/dashboard/ai_predictions/recent", tags=["Frontend"], status_code=200)
async def get_recent_ai_predictions(token: str = Depends(oauth2_scheme)):
    # Implement logic to get recent AI predictions
    return {"predictions": "Recent AI predictions"}

@app.get("/api/dashboard/scaling_logs/recent", tags=["Frontend"], status_code=200)
async def get_recent_scaling_logs(token: str = Depends(oauth2_scheme)):
    # Implement logic to get recent scaling logs
    return {"logs": "Recent scaling logs"}

@app.get("/api/dashboard/alerts/recent", tags=["Frontend"], status_code=200)
async def get_recent_alerts(token: str = Depends(oauth2_scheme)):
    # Implement logic to get recent alerts
    return {"alerts": "Recent alerts"}

if __name__ == '__main__':
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)